requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0
pymilvus==2.6.4
loguru==0.7.3
openai==2.8.1
python-multipart==0.0.20
llama-cpp-python==0.3.16
numpy==2.3.5
pandas==2.3.3
tenacity==9.1.2
elasticsearch==9.2.0
aiohttp==3.13.2
tiktoken==0.12.0
